Add files via upload
A Student Web Portal is an online platform that provides students with access to academic information and resources. It allows students to view course schedules, grades, assignments, and attendance records. The portal also enables communication with teachers, administrators, and peers, and can offer features like fee payment, event updates, and access to study materials. It serves as a central hub for students to manage their academic and extracurricular activities.